+++ src/constants/columns.tsx
@@ -41,6 +41,8 @@ export const hasDetailPageAsPopup = {
     marketSegment: routes?.marketSegment?.path,
     productCategory: routes?.productCategory?.path
 }
+export const popupResources = ["Product Category", "Market Segment", "Warehouse"]
+
 export const disabledColumns = {
     [routes.rentalManagementDetail.title]: [],
     [routes.deliveryTicketDetail.title]: [],
@@ -179,19 +181,27 @@ export const getColumnData = (title, field, detailScreenRoute = null, hasPopup =
                 isForPopup = true
                 pathName = `${hasDetailPageAsPopup[joinedFieldName]}`
             }
+            else if (field?.lookupResource && popupResources.indexOf(field?.lookupResource) >= 0) {
+                pathName = routes[`${camelCase(field?.lookupResource)}`]?.path ?? ""
+                isForPopup = true
+            }
             else {
                 pathName = detailPagePath[joinedFieldName] ? detailPagePath[joinedFieldName] :
                     field?.lookupResource && routes[`${camelCase(field?.lookupResource)}Detail`]?.path ?
                         routes[`${camelCase(field?.lookupResource)}Detail`]?.path :
                         routes[joinedFieldName]?.path ? routes[joinedFieldName]?.path :
                             routes[`${joinedFieldName}Detail`]?.path ? routes[`${joinedFieldName}Detail`]?.path : ""
+
             }
 
             return {
                 columnData: {
                     ...commonFieldData,
                     cellRenderer: "linkRenderer",
-                    cellRendererParams: { "pathName": pathName, "property": joinedFieldName + 'Id', isForPopup: isForPopup, "more": `rest${joinedFieldName}` }
+                    cellRendererParams: {
+                        "pathName": pathName, "property": joinedFieldName + 'Id',
+                        isForPopup: isForPopup, "more": `rest${joinedFieldName}`
+                    }
                 },
                 rendererName: 'linkRenderer',
             }